+++ docs/shared/iframe/iframeResizer-screener.js
@@ -1,3 +1,3 @@
 iFrameResize({
-    log: true
+    log: false
 }, '#snap-iframe')